requirements.txt
Django==3.0.8
djangorestframework==3.11.0
pandas==1.1.2
openpyxl==3.0.5
rest-pandas==1.1.0